refactor: clean up for consistency
