+++ docs/_next/static/chunks/pages/index-26cc2f13d26ed493.js
@@ -1,0 +1,1 @@
+(self.webpackChunk_N_E=self.webpackChunk_N_E||[]).push([[405],{8312:function(n,t,e){(window.__NEXT_P=window.__NEXT_P||[]).push(["/",function(){return e(8321)}])},8321:function(n,t,e){"use strict";e.r(t),e.d(t,{default:function(){return I}});var i=e(2729),r=e(5893),a=e(3012),l=e(5367),o=e(7294);function c(){let n=(0,i._)(["\n  width: 24px;\n  height: 24px;\n  opacity: 0;\n  transition: opacity 0.2s linear;\n\n  &.loaded {\n    opacity: 1;\n  }\n"]);return c=function(){return n},n}let s=a.ZP.svg(c());function p(n){let{className:t,src:e}=n,[i,a]=(0,o.useState)(!1);return(0,o.useEffect)(()=>{a(!0)},[]),(0,r.jsx)(s,{className:"icon".concat(i?" loaded":"").concat(t?" ".concat(t):""),children:(0,r.jsx)("use",{href:e})})}function x(){let n=(0,i._)(["\n  scroll-margin-top: 80px;\n  margin-top: 60px;\n\n  ."," {\n    color: var(",");\n  }\n\n  .desc {\n    margin-top: 48px;\n    white-space: pre-line;\n    text-align: justify;\n  }\n\n  .desc {\n    margin-bottom: 24px;\n  }\n\n  .link {\n    display: inline-flex;\n    align-items: center;\n    color: var(",");\n    text-decoration: none;\n    transition: color 0.2s ease-in-out;\n\n    &.is-last {\n      margin-left: 16px;\n    }\n\n    &:hover {\n      color: var(",");\n    }\n\n    @media "," {\n      &:hover {\n        color: inherit;\n      }\n\n      &:active {\n        color: var(",");\n      }\n    }\n\n    .icon {\n      width: 20px;\n      height: 20px;\n      margin-right: 4px;\n    }\n  }\n\n  @media "," {\n    margin-top: 24px;\n    padding: 0 20px;\n    text-align: center;\n\n    .desc {\n      text-align: center;\n    }\n  }\n"]);return x=function(){return n},n}let d=a.ZP.section(x(),l.Dx.h2,l.r$.secondary,l.r$.accent,l.r$.accentEmp,l.Pr.belowTablet,l.r$.accentEmp,l.Pr.belowTablet);function u(){return(0,r.jsxs)(d,{id:l.wV,children:[(0,r.jsx)("h1",{className:l.Dx.h1,children:l.hS.title}),(0,r.jsx)("h2",{className:l.Dx.h2,children:l.hS.subtitle}),(0,r.jsx)("p",{className:"".concat(l.Dx.body1," desc"),children:l.hS.desc}),(0,r.jsxs)("a",{className:"".concat(l.Dx.subtitle2," link"),href:l.hS.linkedin,target:"_blank",rel:"noopener noreferrer",children:[(0,r.jsx)(p,{src:l.Wm.linkedin})," LinkedIn"]}),(0,r.jsxs)("a",{className:"".concat(l.Dx.subtitle2," link is-last"),href:l.hS.github,target:"_blank",rel:"noopener noreferrer",children:[(0,r.jsx)(p,{src:l.Wm.github})," Github"]})]})}function m(){let n=(0,i._)(["\n  background-color: var(",");\n  color: var(",");\n  border-radius: 48px;\n  padding: 4px 8px;\n"]);return m=function(){return n},n}let g=a.ZP.div(m(),l.r$.secondary,l.r$.neutral);function h(n){let{children:t,className:e}=n;return(0,r.jsx)(g,{className:"".concat(l.Dx.caption2).concat(e?" ".concat(e):""),children:t})}function f(){let n=(0,i._)(["\n  margin: 16px 0 32px;\n  padding-right: 64px;\n\n  .subtitle {\n    display: block;\n    color: var(",");\n    margin-top: 4px;\n  }\n\n  .desc {\n    margin: 12px 0;\n    text-align: justify;\n    white-space: pre-line;\n  }\n\n  .skill-wrapper {\n    display: flex;\n    flex-wrap: wrap;\n    gap: 4px 8px;\n  }\n\n  .subtitle,\n  .desc,\n  .skill-wrapper {\n    padding-left: ","px;\n\n    @media "," {\n      padding-left: 0;\n    }\n  }\n\n  @media "," {\n    margin: 16px 8px 40px;\n    padding-right: 0;\n\n    .subtitle {\n      text-align: center;\n      margin-top: 8px;\n    }\n  }\n"]);return f=function(){return n},n}function b(){let n=(0,i._)(["\n  display: flex;\n  flex-wrap: wrap;\n  align-items: center;\n\n  .icon {\n    flex-shrink: 0;\n    margin-right: ","px;\n    width: ","px;\n    height: ","px;\n  }\n\n  .title {\n    display: inline-block;\n  }\n\n  .title-caption {\n    color: var(",");\n    margin-left: 8px;\n  }\n\n  @media "," {\n    justify-content: center;\n\n    .icon {\n      flex-basis: 100%;\n      margin-right: 0;\n      width: 40px;\n      height: 40px;\n    }\n\n    .title-caption {\n      flex-basis: 100%;\n      text-align: center;\n      margin-left: 0;\n    }\n  }\n"]);return b=function(){return n},n}let j=a.ZP.div(f(),l.r$.secondary,52,l.Pr.belowTablet,l.Pr.belowTablet),w=a.ZP.div(b(),20,32,32,l.r$.lighten,l.Pr.belowTablet);function v(n){let{desc:t,icon:e,skills:i,subtitle:a,title:c,titleCaption:s}=n,x=(0,o.useMemo)(()=>i.map(n=>(0,r.jsx)(h,{children:n},n)),[i]);return(0,r.jsxs)(j,{children:[(0,r.jsxs)(w,{children:[(0,r.jsx)(p,{src:e}),(0,r.jsx)("h3",{className:"".concat(l.Dx.h3," title"),children:c}),(0,r.jsxs)("span",{className:"".concat(l.Dx.caption1," title-caption"),children:["(",s,")"]})]}),(0,r.jsx)("span",{className:"".concat(l.Dx.subtitle1," subtitle"),children:a}),(0,r.jsx)("p",{className:"".concat(l.Dx.body2," desc"),children:t}),(0,r.jsx)("div",{className:"skill-wrapper",children:x})]})}function y(){let n=(0,i._)(["\n  scroll-margin-top: 96px;\n  margin-top: 72px;\n\n  > .title {\n    margin-bottom: 36px;\n  }\n\n  @media "," {\n    padding: 0 16px;\n\n    > .title {\n      text-align: center;\n    }\n  }\n"]);return y=function(){return n},n}let k=a.ZP.section(y(),l.Pr.belowTablet);function N(){let n=(0,o.useMemo)(()=>l.VS.data.map(n=>(0,r.jsx)(v,{desc:n.desc,icon:n.icon,skills:n.skills,subtitle:n.subtitle,title:n.title,titleCaption:n.titleCaption},"".concat(n.title).concat(n.titleCaption))),[]);return(0,r.jsxs)(k,{id:l.Eg,children:[(0,r.jsx)("span",{className:"".concat(l.Dx.h2," title"),children:l.VS.title}),n]})}function _(){let n=(0,i._)(["\n  color: var(",");\n  margin: 96px 16px 32px;\n  text-align: center;\n"]);return _=function(){return n},n}let E=a.ZP.footer(_(),l.r$.lighten);function P(){return(0,r.jsxs)(E,{className:l.Dx.body3,children:["Statically generated by Next.js and hosted on Github Page.",(0,r.jsx)("br",{}),"With help from styled-components and Heroicons."]})}var D=e(3766);function W(){let n=(0,i._)(["\n  ","\n  position: sticky;\n  top: 0;\n  margin: 32px auto 0;\n  padding: 12px 0;\n\n  &::after {\n    content: '';\n    width: 100%;\n    height: 110px;\n    background: linear-gradient(\n      180deg,\n      var(",") 50%,\n      hsla(0, 0%, 100%, 0)\n    );\n    position: absolute;\n    z-index: 0;\n    top: 0;\n    left: 0;\n  }\n\n  .wrapper {\n    display: flex;\n    align-items: center;\n    height: 40px;\n    position: relative;\n    z-index: 1;\n  }\n\n  .logo {\n    display: flex;\n    align-items: center;\n    flex-grow: 1;\n    color: var(",");\n\n    .icon {\n      margin-right: 4px;\n      animation-name: spin;\n      animation-duration: 2.5s;\n      animation-iteration-count: infinite;\n      animation-timing-function: linear;\n\n      @keyframes spin {\n        from {\n          transform: rotate(0deg);\n        }\n        to {\n          transform: rotate(360deg);\n        }\n      }\n    }\n  }\n\n  .links,\n  .toggler {\n    color: var(",");\n  }\n\n  .toggler {\n    display: flex;\n    align-items: center;\n    margin-left: 20px;\n    text-align: right;\n\n    .icon {\n      margin-right: 4px;\n    }\n  }\n\n  .btn {\n    background-color: transparent;\n    border: 0;\n    display: inline-flex;\n    align-items: center;\n    color: var(",");\n    cursor: pointer;\n    padding: 8px;\n    border-radius: 8px;\n    transition: color 0.2s ease-in-out;\n    text-decoration: none;\n\n    &:hover {\n      color: var(",");\n    }\n\n    @media "," {\n      &:hover {\n        color: inherit;\n      }\n\n      &:active {\n        color: var(",");\n      }\n    }\n  }\n\n  @media "," {\n    margin-top: 0;\n\n    .wrapper {\n      padding: 0 16px;\n    }\n\n    .links {\n      margin-right: 0;\n    }\n\n    .toggler {\n      margin-left: 0;\n      flex-grow: 1;\n      justify-content: flex-end;\n\n      .icon {\n        margin-right: 0;\n      }\n    }\n\n    .nav-label {\n      display: none;\n    }\n  }\n"]);return W=function(){return n},n}let M=a.ZP.nav(W(),l.wf,l.r$.neutral,l.r$.secondary,l.r$.accent,l.r$.accent,l.r$.accentEmp,l.Pr.belowTablet,l.r$.accentEmp,l.Pr.belowTablet);function $(n){D.o.getState().toggleIsDarkMode();let t=null==n?void 0:n.target;if(t){var e,i,r;t.blur(),null===(e=t.parentElement)||void 0===e||e.blur(),null===(i=t.parentElement)||void 0===i||null===(r=i.parentElement)||void 0===r||r.blur()}}function T(){let[n,t]=(0,o.useState)(!1),e=(0,D.o)(n=>n.isDarkMode);return(0,o.useEffect)(()=>{t(e)},[e]),(0,r.jsx)(M,{children:(0,r.jsxs)("ul",{className:"wrapper",children:[(0,r.jsxs)("li",{className:"".concat(l.Dx.h3," logo"),children:[(0,r.jsx)(p,{src:l.Wm.cog}),(0,r.jsx)("span",{className:"nav-label",children:"Gear"})]}),(0,r.jsxs)("li",{className:"".concat(l.Dx.subtitle2," links"),children:[(0,r.jsx)("a",{className:"btn",href:"#".concat(l.wV),children:"ABOUT"}),"\xa0/\xa0",(0,r.jsx)("a",{className:"btn",href:"#".concat(l.Eg),children:"EXPERIENCE"})]}),(0,r.jsx)("li",{className:"toggler",children:(0,r.jsxs)("button",{className:"btn",onClick:$,type:"button",children:[(0,r.jsx)(p,{src:n?l.Wm.moon:l.Wm.light}),(0,r.jsx)("span",{className:"".concat(l.Dx.subtitle2," nav-label"),children:n?"DARK":"LIGHT"})]})})]})})}var S=e(9008),Z=e.n(S);function C(){let n=(0,i._)(["\n  ","\n"]);return C=function(){return n},n}let G=a.zo.div(C(),l.wf);function I(){return(0,r.jsxs)(r.Fragment,{children:[(0,r.jsxs)(Z(),{children:[(0,r.jsx)("meta",{name:"viewport",content:"width=device-width, initial-scale=1.0"}),(0,r.jsx)("title",{children:l.MW.title}),(0,r.jsx)("meta",{name:"title",content:l.MW.title}),(0,r.jsx)("meta",{name:"description",content:l.MW.desc}),(0,r.jsx)("meta",{property:"og:type",content:"website"}),(0,r.jsx)("meta",{property:"og:url",content:l.MW.url}),(0,r.jsx)("meta",{property:"og:title",content:l.MW.title}),(0,r.jsx)("meta",{property:"og:description",content:l.MW.desc}),(0,r.jsx)("meta",{property:"og:image",content:l.MW.image}),(0,r.jsx)("meta",{property:"twitter:card",content:"summary_large_image"}),(0,r.jsx)("meta",{property:"twitter:url",content:l.MW.url}),(0,r.jsx)("meta",{property:"twitter:title",content:l.MW.title}),(0,r.jsx)("meta",{property:"twitter:description",content:l.MW.desc}),(0,r.jsx)("meta",{property:"twitter:image",content:l.MW.image}),(0,r.jsx)("link",{rel:"canonical",href:l.MW.url})]}),(0,r.jsx)(T,{}),(0,r.jsxs)(G,{role:"main",children:[(0,r.jsx)(u,{}),(0,r.jsx)(N,{})]}),(0,r.jsx)(P,{})]})}},9008:function(n,t,e){n.exports=e(2636)}},function(n){n.O(0,[774,888,179],function(){return n(n.s=8312)}),_N_E=n.O()}]);
