--- docs/_next/static/chunks/pages/index-f97daec5d2711bdb.js
@@ -1,1 +1,0 @@
-(self.webpackChunk_N_E=self.webpackChunk_N_E||[]).push([[405],{8312:function(e,n,t){(window.__NEXT_P=window.__NEXT_P||[]).push(["/",function(){return t(7555)}])},7555:function(e,n,t){"use strict";t.r(n),t.d(n,{default:function(){return U}});var a=t(2729),i=t(5893),r=t(3012),o=t(6206);let l="about";var s=t(1788);let c={title:"Gear - Chalermkried Sudjanakobkul",desc:"A Front-End Engineer who likes, create, and optimize websites, problem-solving, helping people, and appreciate arts.",image:"https://chalermkried.github.io/assets/cover.jpg",url:"https://chalermkried.github.io"},d={title:"Chalermkried Sudjanakobkul",subtitle:"Front-End Engineer",desc:"Hi, I'm Gear!\n\nI like to create and optimize websites, make them fast, easy to use, minimal efforts needed, an enjoyable experience for the users. I care deeply about the UI and UX, make use of organic traffics with SEO, and last but not least, accessibility. With all those preferred results, I always look for the best tools to help in the development, new tech stacks, and alternatives.\n  \nI also like to help people and solve problems, those challenges are sometimes enjoyable!\n  \nInterests: music, stories, football, philosophy, and awesome people!",github:"https://github.com/chalermkried/",linkedin:"https://www.linkedin.com/in/gear-chalermkried/"},p={title:"Experience",data:[{desc:"Back to a leaner working environment and resuming a more-frontend focused role. With my position as a senior member of the team, my work scope expands to more than just coding and delivering features.\n      \n      The applications that I'm working on this time are more logic-heavy and involve many connected parts, state management, and data manipulation is crucial. Our team's works also involve the process of migrating older codebase to a newer stack which requires regular collaboration and communication between engineers.\n      \n      Software paradigms, concepts, and architecture design play a vital role to help the company, team, and application scaling.",icon:s.W.terminal,skills:["CircleCI","Semantic Release","React Query","Clean Architecture","Python","Django"],subtitle:"CREA",title:"Senior Frontend Engineer",titleCaption:"Dec 2021 - Present"},{desc:"Worked in a financial-related, corporate firm, for the first time, here I have the chance to work on many projects, both customer-facing and internal products. By working on stacks such as Front-end, BFF, Gateway, APIs, and Back-end.\n      \n      Under the technology team, there are many new tools and different service providers that I have learned to use. Also more hands-on experience in product and application building from the bottom to the top.\n            \n      The works here are pretty broads. I have a lot more opportunities to work in the DevOps part of the product cycle than I had before. My daily tasks as a developer here are less on the Front-end parts. But spread out into many stacks with more focus on the overview of the applications.\n      \n      Additionally, as a financial firm, the applications have a lot of details and focus on security, legal/consent, and data integrity.",icon:s.W.terminal,skills:["Backends for Frontends","APIs","Gateway","Oauth2 & JWT","Sequelize","Kubernetes","Helm","Docker","Azure","Groovy","DevSecOps"],subtitle:"Prudential Life Assurance (Thailand) PCL",title:"Full-Stack Developer",titleCaption:"Nov 2020 - Dec 2021"},{desc:"With the new role, I carry over most of the tasks from the previous role. But now as a Senior, I have the chance to help and teach other members of the team more. Also more responsibility to make decisions in the meetings with other teams.\n\nAfter a few months, due to the circumstances, I got to become the web team lead, internally. With this, it also meant my responsibility now covers more on the higher levels tasks such as team management, web monitoring, infrastructure monitoring, release management, documentation, and managing the team of outsource-developers.",icon:s.W.terminal,skills:["Github","Github Action","CI/CD","Nginx","A/B Testing","Segment","Sentry","Datadog","Team Management","Task Delegation","AWS","CDN","a11y","Stress Testing"],subtitle:"Pomelofashion",title:"Senior Front-End Engineer",titleCaption:"Feb 2020 - Nov 2020"},{desc:"Worked as part of the Front-End web team, to maintain, implement new features, and support campaigns with the newly released Pomelo web (at that time).\n\nThe Pomelo web is the main product to focus on, I closely worked with the UI/UX team to implement and improve the website following the mockups with prioritization from the Product team. Also correlated with the QA team to resolve issues and bugs. There are also other web projects such as Pomelo Screens and Internal Dashboards that I became the lead developer at the latter part of this time.\n      \nDuring this time, I also had the chance to work in the LEAD program, an internal employee development program. With the program, I got to learn how to work as an assistant product manager for Pomelo Partner project.",icon:s.W.terminal,skills:["React","React Hook","SEO","Next.js","Server Side Rendering","Jenkins","Redux","Segment","Session Monitoring","GTM & GA","Product Management"],subtitle:"Pomelofashion",title:"Front-End Engineer",titleCaption:"Aug 2018 - Jan 2020"},{desc:"Worked as a member of the Front-End web development team, in a multicultural software house company with a full team of product/software development consists of Designers (UI/UX), Products (PM/PO), Testers (QA), and Backend.\n\nHad the chance to work in many projects, the first project was the project that provided a good learning experience about the latest web technology and how does software development really like. One of the projects was a large one with a tight deadline going into MVP development phase, and we also provided an ongoing support after it's launched.",icon:s.W.terminal,skills:["TypeScript","Git","Bitbucket","JIRA","Scrum","Agile","REST","Angular","RxJS","NgRx"],subtitle:"7 Peaks Software",title:"Junior Front-End Developer",titleCaption:"Jun 2017 -  Aug 2018"},{desc:"Studied in Computer Science, learned about Data Structure, Algorithm, Mathematics, Statistics, and various topics related to the fundamentals of programming and problem-solving.\n\nYielded many full-stack web projects, and experienced in other programming languages and product development, such as C, C#, Swift, Visual Basic, mobile development, game development with Unity, and product development with system analysis.",icon:s.W.library,skills:["HTML","CSS","JavaScript","jQuery","PHP","MySQL","Oracle","Apache","Monolithic"],subtitle:"Srinakharinwirot University",title:"Student",titleCaption:"2014 - 2017"}]};var m=t(7294);function h(){let e=(0,a._)(["\n  width: 24px;\n  height: 24px;\n  opacity: 0;\n  transition: opacity 0.2s linear;\n\n  &.loaded {\n    opacity: 1;\n  }\n"]);return h=function(){return e},e}let u=r.ZP.svg(h());function g(e){let{className:n,src:t}=e,[a,r]=(0,m.useState)(!1);return(0,m.useEffect)(()=>{r(!0)},[]),(0,i.jsx)(u,{className:"icon".concat(a?" loaded":"").concat(n?" ".concat(n):""),children:(0,i.jsx)("use",{href:t})})}function x(){let e=(0,a._)(["\n  scroll-margin-top: 80px;\n  margin-top: 60px;\n\n  ."," {\n    color: var(",");\n  }\n\n  .desc {\n    margin-top: 48px;\n    white-space: pre-line;\n    text-align: justify;\n  }\n\n  .desc {\n    margin-bottom: 24px;\n  }\n\n  .link {\n    display: inline-flex;\n    align-items: center;\n    color: var(",");\n    text-decoration: none;\n    transition: color 0.2s ease-in-out;\n\n    &.is-last {\n      margin-left: 16px;\n    }\n\n    &:hover {\n      color: var(",");\n    }\n\n    @media "," {\n      &:hover {\n        color: inherit;\n      }\n\n      &:active {\n        color: var(",");\n      }\n    }\n\n    .icon {\n      width: 20px;\n      height: 20px;\n      margin-right: 4px;\n    }\n  }\n\n  @media "," {\n    margin-top: 24px;\n    padding: 0 20px;\n    text-align: center;\n\n    .desc {\n      text-align: center;\n    }\n  }\n"]);return x=function(){return e},e}let f=r.ZP.section(x(),o.Dx.h2,o.r$.secondary,o.r$.accent,o.r$.accentEmp,o.Pr.belowTablet,o.r$.accentEmp,o.Pr.belowTablet);function b(){return(0,i.jsxs)(f,{id:l,children:[(0,i.jsx)("h1",{className:o.Dx.h1,children:d.title}),(0,i.jsx)("h2",{className:o.Dx.h2,children:d.subtitle}),(0,i.jsx)("p",{className:"".concat(o.Dx.body1," desc"),children:d.desc}),(0,i.jsxs)("a",{className:"".concat(o.Dx.subtitle2," link"),href:d.linkedin,target:"_blank",rel:"noopener noreferrer",children:[(0,i.jsx)(g,{src:o.Wm.linkedin})," LinkedIn"]}),(0,i.jsxs)("a",{className:"".concat(o.Dx.subtitle2," link is-last"),href:d.github,target:"_blank",rel:"noopener noreferrer",children:[(0,i.jsx)(g,{src:o.Wm.github})," Github"]})]})}function w(){let e=(0,a._)(["\n  background-color: var(",");\n  color: var(",");\n  border-radius: 48px;\n  padding: 4px 8px;\n"]);return w=function(){return e},e}let v=r.ZP.div(w(),o.r$.secondary,o.r$.neutral);function y(e){let{children:n,className:t}=e;return(0,i.jsx)(v,{className:"".concat(o.Dx.caption2).concat(t?" ".concat(t):""),children:n})}function k(){let e=(0,a._)(["\n  margin: 16px 0 32px;\n  padding-right: 64px;\n\n  .subtitle {\n    display: block;\n    color: var(",");\n    margin-top: 4px;\n  }\n\n  .desc {\n    margin: 12px 0;\n    text-align: justify;\n    white-space: pre-line;\n  }\n\n  .skill-wrapper {\n    display: flex;\n    flex-wrap: wrap;\n    gap: 4px 8px;\n  }\n\n  .subtitle,\n  .desc,\n  .skill-wrapper {\n    padding-left: ","px;\n\n    @media "," {\n      padding-left: 0;\n    }\n  }\n\n  @media "," {\n    margin: 16px 8px 40px;\n    padding-right: 0;\n\n    .subtitle {\n      text-align: center;\n      margin-top: 8px;\n    }\n  }\n"]);return k=function(){return e},e}function j(){let e=(0,a._)(["\n  display: flex;\n  flex-wrap: wrap;\n  align-items: center;\n\n  .icon {\n    flex-shrink: 0;\n    margin-right: ","px;\n    width: ","px;\n    height: ","px;\n  }\n\n  .title {\n    display: inline-block;\n  }\n\n  .title-caption {\n    color: var(",");\n    margin-left: 8px;\n  }\n\n  @media "," {\n    justify-content: center;\n\n    .icon {\n      flex-basis: 100%;\n      margin-right: 0;\n      width: 40px;\n      height: 40px;\n    }\n\n    .title-caption {\n      flex-basis: 100%;\n      text-align: center;\n      margin-left: 0;\n    }\n  }\n"]);return j=function(){return e},e}let P=r.ZP.div(k(),o.r$.secondary,52,o.Pr.belowTablet,o.Pr.belowTablet),S=r.ZP.div(j(),20,32,32,o.r$.lighten,o.Pr.belowTablet);function D(e){let{desc:n,icon:t,skills:a,subtitle:r,title:l,titleCaption:s}=e,c=(0,m.useMemo)(()=>a.map(e=>(0,i.jsx)(y,{children:e},e)),[a]);return(0,i.jsxs)(P,{children:[(0,i.jsxs)(S,{children:[(0,i.jsx)(g,{src:t}),(0,i.jsx)("h3",{className:"".concat(o.Dx.h3," title"),children:l}),(0,i.jsxs)("span",{className:"".concat(o.Dx.caption1," title-caption"),children:["(",s,")"]})]}),(0,i.jsx)("span",{className:"".concat(o.Dx.subtitle1," subtitle"),children:r}),(0,i.jsx)("p",{className:"".concat(o.Dx.body2," desc"),children:n}),(0,i.jsx)("div",{className:"skill-wrapper",children:c})]})}function E(){let e=(0,a._)(["\n  scroll-margin-top: 96px;\n  margin-top: 72px;\n\n  > .title {\n    margin-bottom: 36px;\n  }\n\n  @media "," {\n    padding: 0 16px;\n\n    > .title {\n      text-align: center;\n    }\n  }\n"]);return E=function(){return e},e}let N=r.ZP.section(E(),o.Pr.belowTablet);function I(){let e=(0,m.useMemo)(()=>p.data.map(e=>(0,i.jsx)(D,{desc:e.desc,icon:e.icon,skills:e.skills,subtitle:e.subtitle,title:e.title,titleCaption:e.titleCaption},"".concat(e.title).concat(e.titleCaption))),[]);return(0,i.jsxs)(N,{id:"exp",children:[(0,i.jsx)("span",{className:"".concat(o.Dx.h2," title"),children:p.title}),e]})}function A(){let e=(0,a._)(["\n  color: var(",");\n  margin: 96px 16px 32px;\n  text-align: center;\n"]);return A=function(){return e},e}let C=r.ZP.footer(A(),o.r$.lighten);function T(){return(0,i.jsxs)(C,{className:o.Dx.body3,children:["Statically generated by Next.js and hosted on Github Page.",(0,i.jsx)("br",{}),"With help from styled-components and Heroicons."]})}var _=t(3766);function W(){let e=(0,a._)(["\n  ","\n  position: sticky;\n  top: 0;\n  margin: 32px auto 0;\n  padding: 12px 0;\n\n  &::after {\n    content: '';\n    width: 100%;\n    height: 110px;\n    background: linear-gradient(\n      180deg,\n      var(",") 50%,\n      hsla(0, 0%, 100%, 0)\n    );\n    position: absolute;\n    z-index: 0;\n    top: 0;\n    left: 0;\n  }\n\n  .wrapper {\n    display: flex;\n    align-items: center;\n    height: 40px;\n    position: relative;\n    z-index: 1;\n  }\n\n  .logo {\n    display: flex;\n    align-items: center;\n    flex-grow: 1;\n    color: var(",");\n\n    .icon {\n      margin-right: 4px;\n      animation-name: spin;\n      animation-duration: 2.5s;\n      animation-iteration-count: infinite;\n      animation-timing-function: linear;\n\n      @keyframes spin {\n        from {\n          transform: rotate(0deg);\n        }\n        to {\n          transform: rotate(360deg);\n        }\n      }\n    }\n  }\n\n  .links,\n  .toggler {\n    color: var(",");\n  }\n\n  .toggler {\n    display: flex;\n    align-items: center;\n    margin-left: 20px;\n    text-align: right;\n\n    .icon {\n      margin-right: 4px;\n    }\n  }\n\n  .btn {\n    background-color: transparent;\n    border: 0;\n    display: inline-flex;\n    align-items: center;\n    color: var(",");\n    cursor: pointer;\n    padding: 8px;\n    border-radius: 8px;\n    transition: color 0.2s ease-in-out;\n    text-decoration: none;\n\n    &:hover {\n      color: var(",");\n    }\n\n    @media "," {\n      &:hover {\n        color: inherit;\n      }\n\n      &:active {\n        color: var(",");\n      }\n    }\n  }\n\n  @media "," {\n    margin-top: 0;\n\n    .wrapper {\n      padding: 0 16px;\n    }\n\n    .links {\n      margin-right: 0;\n    }\n\n    .toggler {\n      margin-left: 0;\n      flex-grow: 1;\n      justify-content: flex-end;\n\n      .icon {\n        margin-right: 0;\n      }\n    }\n\n    .nav-label {\n      display: none;\n    }\n  }\n"]);return W=function(){return e},e}let F=r.ZP.nav(W(),o.wf,o.r$.neutral,o.r$.secondary,o.r$.accent,o.r$.accent,o.r$.accentEmp,o.Pr.belowTablet,o.r$.accentEmp,o.Pr.belowTablet);function M(e){_.o.getState().toggleIsDarkMode();let n=null==e?void 0:e.target;if(n){var t,a,i;n.blur(),null===(t=n.parentElement)||void 0===t||t.blur(),null===(a=n.parentElement)||void 0===a||null===(i=a.parentElement)||void 0===i||i.blur()}}function $(){let[e,n]=(0,m.useState)(!1),t=(0,_.o)(e=>e.isDarkMode);return(0,m.useEffect)(()=>{n(t)},[t]),(0,i.jsx)(F,{children:(0,i.jsxs)("ul",{className:"wrapper",children:[(0,i.jsxs)("li",{className:"".concat(o.Dx.h3," logo"),children:[(0,i.jsx)(g,{src:o.Wm.cog}),(0,i.jsx)("span",{className:"nav-label",children:"Gear"})]}),(0,i.jsxs)("li",{className:"".concat(o.Dx.subtitle2," links"),children:[(0,i.jsx)("a",{className:"btn",href:"#".concat(l),children:"ABOUT"}),"\xa0/\xa0",(0,i.jsx)("a",{className:"btn",href:"#".concat("exp"),children:"EXPERIENCE"})]}),(0,i.jsx)("li",{className:"toggler",children:(0,i.jsxs)("button",{className:"btn",onClick:M,type:"button",children:[(0,i.jsx)(g,{src:e?o.Wm.moon:o.Wm.light}),(0,i.jsx)("span",{className:"".concat(o.Dx.subtitle2," nav-label"),children:e?"DARK":"LIGHT"})]})})]})})}var G=t(9008),R=t.n(G);function B(){let e=(0,a._)(["\n  ","\n"]);return B=function(){return e},e}let O=r.zo.div(B(),o.wf);function U(){return(0,i.jsxs)(i.Fragment,{children:[(0,i.jsxs)(R(),{children:[(0,i.jsx)("meta",{name:"viewport",content:"width=device-width, initial-scale=1.0"}),(0,i.jsx)("title",{children:c.title}),(0,i.jsx)("meta",{name:"title",content:c.title}),(0,i.jsx)("meta",{name:"description",content:c.desc}),(0,i.jsx)("meta",{property:"og:type",content:"website"}),(0,i.jsx)("meta",{property:"og:url",content:c.url}),(0,i.jsx)("meta",{property:"og:title",content:c.title}),(0,i.jsx)("meta",{property:"og:description",content:c.desc}),(0,i.jsx)("meta",{property:"og:image",content:c.image}),(0,i.jsx)("meta",{property:"twitter:card",content:"summary_large_image"}),(0,i.jsx)("meta",{property:"twitter:url",content:c.url}),(0,i.jsx)("meta",{property:"twitter:title",content:c.title}),(0,i.jsx)("meta",{property:"twitter:description",content:c.desc}),(0,i.jsx)("meta",{property:"twitter:image",content:c.image}),(0,i.jsx)("link",{rel:"canonical",href:c.url})]}),(0,i.jsx)($,{}),(0,i.jsxs)(O,{role:"main",children:[(0,i.jsx)(b,{}),(0,i.jsx)(I,{})]}),(0,i.jsx)(T,{})]})}},9008:function(e,n,t){e.exports=t(2636)}},function(e){e.O(0,[774,888,179],function(){return e(e.s=8312)}),_N_E=e.O()}]);
