+++ src/components/exp/card.tsx
