+++ src/const/index.ts
@@ -1,3 +1,5 @@
+export * from './content'
+export * from './element'
 export * from './icon'
 export * from './storage'
 export * from './style'
