+++ src/pages/index.tsx
@@ -2,8 +2,7 @@ import { About } from '@/components/about'
 import { Exp } from '@/components/exp'
 import { Footer } from '@/components/footer'
 import { TopNav } from '@/components/top-nav'
-import { CONTAINER_STYLE } from '@/const'
-import { META } from '@/const/content'
+import { CONTAINER_STYLE, META } from '@/const'
 import Head from 'next/head'
 import { styled } from 'styled-components'
